Add new project and debugging tools for color contrast
Added a new project "Lesson 3 - color contrast" to the solution file `TeachingAccessibility.sln`. This includes specific configurations for both Debug and Release modes. Updated the global section of the solution file to include build configurations for the new "ColorContrast" project.

Added new HTML, CSS, and JavaScript files to support the development and debugging of the new project:
- `index.html` with basic structure and a script to toggle CSS outlines.
- `debugger.css` to outline all elements on the page for debugging.
- `debugHelper.js` and its minified version `debugHelper.min.js` to manage the inclusion and removal of `debugger.css`.
- Source map file `debugHelper.min.js.map` for the minified JavaScript file.

Added a `readme.md` file with a brief description of the project and a `LICENSE` file with the MIT License.