Enhance accessibility documentation
Added an image to `readme.md` to improve visual reference for accessibility. Introduced a new "Screen Reader Testing" section in `screenReaders.md`, detailing the function and importance of screen readers, best practices for accessibility, and practical guidance for testing without a mouse. Included a link to an external resource for further understanding.